Major refactor done.
This commit represents the state at which a major refactor has been done and
the code has been tested for both argon/xenon and multilayer images. This code
also uses the new wcifs registry values to specify the boot.cim path instead of
using wcitool to add binds for the boot.cim.