Remove unnecessary cim mount cache
Currently we have a map which maintains a mapping of CIM & containerd ID to the volume at
which a CIM is mounted for the given container. This was required before the layer
refactoring work when we needed to get the volume path from the layer cim path. However,
this isn't needed anymore. As of now, this map doesn't provide much value and makes the code a
bit complicated. Moreover, we will need to rewrite some of this code anyway when we do the work
required for handling `shim delete` cleanups properly (https://github.com/containerd/containerd/issues/9727).

Signed-off-by: Amit Barve <ambarve@microsoft.com>